Refactor error handling for BINOSPEC and MMIRS APIs by moving catch_timeout_and_no_endpoint to utils.py
